update start local scripts to work with aws "credential_process" profile

+++ scripts/start_local.sh
@@ -1,4 +1,5 @@
 #!/bin/bash
+set -e
 
 unset VAULT_TOKEN
 export VAULT_ADDR='http://127.0.0.1:8200'
@@ -24,11 +25,28 @@ if [ -z "$ARGO_CLOUDOPS_GIT_AUTH_METHOD" ]; then
     export ARGO_CLOUDOPS_GIT_AUTH_METHOD=https
 fi
 
-# TODO / HACK: Vault was not loading credentials from the default chain.
+# Vault was not loading credentials from the default chain, try to fetch from profile
 if [ -n "${AWS_PROFILE}" ]; then
-    export AWS_ACCESS_KEY_ID=`aws configure get $AWS_PROFILE.aws_access_key_id`
-    export AWS_SECRET_ACCESS_KEY=`aws configure get $AWS_PROFILE.aws_secret_access_key`
-    export AWS_SESSION_TOKEN=`aws configure get $AWS_PROFILE.aws_session_token`
+    CREDS_PROCESS_VALUE=`aws configure get $AWS_PROFILE.credential_process`
+    # export AWS_DEFAULT_REGION=`aws configure get $AWS_PROFILE.region`
+    if [ -n "$CREDS_PROCESS_VALUE" ]; then
+        # profile is using credential_process
+
+        # git account ID first
+        export ACCOUNT_ID=`aws sts get-caller-identity --query Account --output text`
+
+        # parse json from credential_process
+        CREDS_JSON=`$CREDS_PROCESS_VALUE`
+        export AWS_ACCESS_KEY_ID=`echo "$CREDS_JSON" | jq ."AccessKeyId"`
+        export AWS_SECRET_ACCESS_KEY=`echo "$CREDS_JSON" | jq ."SecretAccessKey"`
+        export AWS_SESSION_TOKEN=`echo "$CREDS_JSON" | jq ."SessionToken"`
+    else
+        # profile isn't using credential_process, get values from profile config
+        export AWS_ACCESS_KEY_ID=`aws configure get $AWS_PROFILE.aws_access_key_id`
+        export AWS_SECRET_ACCESS_KEY=`aws configure get $AWS_PROFILE.aws_secret_access_key`
+        export AWS_SESSION_TOKEN=`aws configure get $AWS_PROFILE.aws_session_token`
+        export ACCOUNT_ID=`aws sts get-caller-identity --query Account --output text`
+    fi
 fi
 
 if [ -z "$ARGO_CLOUDOPS_CONFIG" ]; then
@@ -50,9 +68,14 @@ if [ -z "$AWS_SESSION_TOKEN" ]; then
     exit 1
 fi
 
-set -e
+if [ -z "$ACCOUNT_ID" ]; then
+    export ACCOUNT_ID=`aws sts get-caller-identity --query Account --output text`
+fi
+if [ -z "$ACCOUNT_ID" ]; then
+    echo "Account ID not found!"
+    exit 1
+fi
 
-export ACCOUNT_ID=`aws sts get-caller-identity --query Account --output text`
 echo "Starting with credentials in AWS account '$ACCOUNT_ID'."
 
 pkill -9 vault && true
